refactor(device-share): migrate DeviceShareValidator to DeviceShareGuard pattern
Migrate DeviceShareValidator to DeviceShareGuard applying the guard clause pattern
to reduce ambiguity.

- Renamed DeviceShareValidator to DeviceShareGuard
- Updated imports and usages in CreateDeviceShareCodeAction,
GetDeviceByShareCodeAction
